+++ main.go
@@ -83,6 +83,7 @@ func main() {
 	commands.Init_Diff_Command(rootCmd, &app)
 	commands.Init_Down_Command(rootCmd, &app)
 	commands.Init_Exec_Command(rootCmd, &app)
+	commands.Init_Generate_Command(rootCmd, &app)
 	commands.Init_Import_Command(rootCmd, &app)
 	commands.Init_Init_Command(rootCmd, &app)
 	commands.Init_Install_Command(rootCmd, &app)